update type int -> decimal in money athributes

+++ backend/src/models/project_model.py
@@ -5,6 +5,7 @@ from core.database import Base
 from core.enums import ProjectCategoryEnum, ProjectStatusEnum
 from sqlalchemy import DateTime, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship, validates
+from sqlalchemy.types import DECIMAL
 
 if TYPE_CHECKING:
     from models.contract_model import Contract  # noqa: TC004
@@ -17,7 +18,7 @@ class Project(Base):
 
     title: Mapped[str] = mapped_column(String(255), nullable=False, index=True)
     description: Mapped[str] = mapped_column(nullable=False)
-    budget: Mapped[int] = mapped_column(nullable=False)
+    budget: Mapped[DECIMAL] = mapped_column(nullable=False)
     deadline: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
     category: Mapped[ProjectCategoryEnum] = mapped_column(nullable=False)
